requirements.txt
asgiref==3.8.1
Django==5.2
django-filter==25.1
psycopg==3.2.6
psycopg-binary==3.2.6
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
